--- include/aquarius/any_io_executor.hpp
@@ -1,8 +1,0 @@
-#pragma once
-#include <aquarius/detail/config.hpp>
-#include <boost/asio/any_io_executor.hpp>
-
-namespace aquarius
-{
-	using boost::asio::any_io_executor;
-}
