+++ include/aquarius/executor/tcp_server_session.hpp
@@ -1,8 +1,8 @@
 #pragma once
-#include <aquarius/basic_session.hpp>
-#include <aquarius/protocol/tcp.hpp>
+#include <aquarius/basic_tcp_session.hpp>
+#include <virgo.hpp>
 
 namespace aquarius
 {
-	using tcp_server_session = basic_session<true, tcp>;
+	using tcp_server_session = basic_tcp_session<true, virgo::tcp::detail::layer_header>;
 }
