+++ include/aquarius/protocol/http.hpp
@@ -1,7 +1,6 @@
 #pragma once
-#include <aquarius/awaitable.hpp>
-#include <aquarius/detail/handler_router.hpp>
-#include <aquarius/error_code.hpp>
+#include <aquarius/basic_session.hpp>
+#include <virgo.hpp>
 #include <ranges>
 #include <virgo.hpp>
 
@@ -11,6 +10,8 @@ namespace aquarius
 	class http
 	{
 		constexpr static std::size_t max_http_length = 8192;
+	public:
+		using header = virgo::http::detail::basic_header<Server>;
 
 	public:
 		using header = virgo::http::detail::basic_header<Server>;
